// src/models/User.ts
import mongoose from 'mongoose';
import uniqueValidator from 'mongoose-unique-validator';

// Id field automatically generated by mongoose.
const userSchema = new mongoose.Schema({
  email: { type: String, required: true, unique: true },
  password: { type: String, required: true }
});

userSchema.plugin(uniqueValidator);

const User = mongoose.model('User', userSchema);

export default User;